Reimplemented roleCard cross fade with divs with toggling opacity New Sourcing role card also created in StudentApply.tsx
RoleCard.tsx updated with different HTML structure now. There are two
divs within cardContent class, cardFront and cardBack. Hovering over one
or the other toggles the opacity. This implementation makes it easier
for the text in backCard div to sit within.

cardFront and cardBack take a similar role to former roleCardTextMain
and roleCardTextSub, but instead they comprise the entire "front" and
"back" of the card, eliminating redundant hover rules/opacity controls
for multiple elements.

Only setback may be that the background images to the cards are now done
inline, instead of with <img> tag. Should not change reverse toggle
behaviour, however