Partial fix of issue of bottom of home page
Cleaned up CSS file a bit and made the two divs holding the info for the
Students and Nonprofits sections the same size in the div. Still need to
figure out how to force the buttons in both divs to stick to the bottom
of the div, making them always aligned.