Fix: make Home Page buttons vertically aligned
Updated the component for the lower part of the Home Page and the CSS
styling for the component to make Home Page buttons vertically aligned
in desktop layout:
- Put text and buttons in separate divs in order to make sure buttons
stick to bottom of the student and nonprofits containers.
- Added flex-direction and flex-grow attributes to CSS file and made the
textHolder class always take up the full height of the div minus the
height of the button div to make sure that button does not overlap with
text elements.
- Also cleaned up CSS a little for the mobile view to make margins and
padding easier to see.