Update styling on top of home page
Fixed issue where the text in the top of the home page would overlap
with the navigation bar when zoomed in / in mobile view.

Added styling to fix issue with a horizontal scoll bar showing up
on the home page. There is still an issue with the bottom of the blue
wave background image getting cut off that is unresolved.